Windows: test symlinks and unix paths
This commit adds tests for symlinked directories and files with long
paths on Windows.  This test is not being added due to any discovered
flaws in the current path handling, but instead is meant to improve
coverage.

One thing to note is that it appears RemoveAll may not remove file
links.

Additionally, an explicit test for unix paths was added (this was
previously tested) and a test for mixed Windows and unix paths.